Module acter::api

Modules

api

Structs

Account

ActerAppSettings

ActerAppSettingsBuilder

ActerPin

Attachment

AttachmentDraft

AttachmentsManager

CalendarEvent

CalendarEventDraft

CalendarEventUpdateBuilder

Client

ClientStateBuilder

Builder for ClientState.

Colorize

Customize the color scheme

ColorizeBuilder

Comment

CommentDraft

CommentsManager

Convo

CreateConvoSettings

CreateConvoSettingsBuilder

Builder for CreateConvoSettings.

CreateSpaceSettings

CreateSpaceSettingsBuilder

Builder for CreateSpaceSettings.

DeviceChangedEvent

DeviceNewEvent

DeviceRecord

EfkDuration

A Duration type to represent a span of time, typically used for system timeouts.

EventSendState

HistoryLoadState

Invitation

JoinRuleBuilder

MediaSource

Member

NewsEntry

NewsEntryDraft

NewsEntryUpdateBuilder

NewsSlide

NewsSlideDraft

Notification

NotificationItem

NotificationListResult

NotificationSettings

ObjRef

An object reference is a link within the application to a specific object with an optional flag to explain how to embed said object. These may be interactive elements when rendered on the view.

ObjRefBuilder

OptionBuffer

OptionRsvpStatus

OptionString

PinDraft

PinUpdateBuilder

PublicSearchResult

PublicSearchResultItem

Pusher

RUNTIME

ReactionRecord

ReceiptEvent

ReceiptRecord

ReceiptThread

RefDetailsBuilder

Room

RoomEventItem

RoomMessage

RoomPowerLevels

RoomProfile

RoomVirtualItem

Rsvp

RsvpDraft

RsvpManager

SessionManager

SimpleSettingWithTurnOff

SimpleSettingWithTurnOffBuilder

Builder for SimpleSettingWithTurnOff.

Space

SpaceHierarchyListResult

SpaceHierarchyRoomInfo

SpaceRelation

SpaceRelations

SuperInviteToken

SuperInvites

SuperInvitesTokenUpdateBuilder

SyncState

Tag

Task

TaskDraft

TaskList

TaskListDraft

TaskListUpdateBuilder

TaskUpdateBuilder

TasksSettings

TasksSettingsBuilder

Builder for TasksSettings.

TextMessageContent

ThreePidManager

ThumbnailInfo

ThumbnailSize

TimelineStream

TypingEvent

VerificationEmoji

VerificationEvent

Enums

MemberPermission

MembershipStatus

MsgContent

MsgContentDraft

NewsContent

RefDetails

RelationTargetType

RsvpStatus

RSVP status

UserProfile

Traits

ActerModel

Functions

__hello_world

destroy_local_data

duration_from_secs

guest_client

login_new_client

login_with_token

new_calendar_event_ref_builder

new_colorize_builder

new_convo_settings_builder

new_join_rule_builder

new_link_ref_builder

new_obj_ref_builder

new_space_settings_builder

new_task_list_ref_builder

new_task_ref_builder

new_thumb_size

parse_markdown

register_with_token

set_proxy

Type Aliases

ConvoDiff

DeviceId

EventId

EventsSettings

MxcUri

NewsSettings

PinsSettings

RoomId

SpaceDiff

TimelineDiff

UserId

UtcDateTime

Default UTC DateTime Object