Trait acter::api::ActerModel

pub trait ActerModel: Debug {
    // Required methods
    fn indizes(&self, user_id: &UserId) -> Vec<String>;
    fn event_id(&self) -> &EventId;
    async fn execute(self, store: &Store) -> Result<Vec<String>, Error>;

    // Provided methods
    fn belongs_to(&self) -> Option<Vec<String>> { ... }
    fn capabilities(&self) -> &[Capability] { ... }
    fn transition(&mut self, model: &AnyActerModel) -> Result<bool, Error> { ... }
    async fn redact(
        &self,
        store: &Store,
        redaction_model: RedactedActerModel
    ) -> Result<Vec<String>, Error> { ... }
}

Required Methods

fn indizes(&self, user_id: &UserId) -> Vec<String>

fn event_id(&self) -> &EventId

The key to store this model under

async fn execute(self, store: &Store) -> Result<Vec<String>, Error>

The execution to run when this model is found.

Provided Methods

fn belongs_to(&self) -> Option<Vec<String>>

The models to inform about this model as it belongs to that

fn capabilities(&self) -> &[Capability]

activate to enable commenting support for this type of model

fn transition(&mut self, model: &AnyActerModel) -> Result<bool, Error>

handle transition from an external Item upon us

async fn redact( &self, store: &Store, redaction_model: RedactedActerModel ) -> Result<Vec<String>, Error>

The execution to run when this model is found.

Object Safety

This trait is not object safe.